room: move index search functions to libtrx
