+++ src/libtrx/game/rooms/common.c
@@ -1,6 +1,7 @@
 #include "game/rooms/common.h"
 
 #include "debug.h"
+#include "game/const.h"
 #include "game/gamebuf.h"
 #include "game/items.h"
 #include "game/rooms/const.h"
@@ -184,4 +185,32 @@ int32_t Room_GetAdjoiningRooms(
     }
 
     return count;
+}
+
+int16_t Room_GetIndexFromPos(const int32_t x, const int32_t y, const int32_t z)
+{
+    // TODO: merge this to Room_FindByPos!
+    const int32_t room_num = Room_FindByPos(x, y, z);
+    if (room_num == NO_ROOM_NEG) {
+        return NO_ROOM;
+    }
+    return room_num;
+}
+
+int32_t Room_FindByPos(const int32_t x, const int32_t y, const int32_t z)
+{
+    for (int32_t i = 0; i < Room_GetTotalCount(); i++) {
+        const ROOM *const room = Room_Get(i);
+        const int32_t x1 = room->pos.x + WALL_L;
+        const int32_t x2 = room->pos.x + (room->size.x - 1) * WALL_L;
+        const int32_t y1 = room->max_ceiling;
+        const int32_t y2 = room->min_floor;
+        const int32_t z1 = room->pos.z + WALL_L;
+        const int32_t z2 = room->pos.z + (room->size.z - 1) * WALL_L;
+        if (x >= x1 && x < x2 && y >= y1 && y <= y2 && z >= z1 && z < z2) {
+            return i;
+        }
+    }
+
+    return NO_ROOM_NEG;
 }
