+++ src/tr1/game/room.h
@@ -23,7 +23,6 @@ SECTOR *Room_GetPitSector(const SECTOR *sector, int32_t x, int32_t z);
 int16_t Room_GetCeiling(const SECTOR *sector, int32_t x, int32_t y, int32_t z);
 int16_t Room_GetHeight(const SECTOR *sector, int32_t x, int32_t y, int32_t z);
 int16_t Room_GetWaterHeight(int32_t x, int32_t y, int32_t z, int16_t room_num);
-int16_t Room_GetIndexFromPos(int32_t x, int32_t y, int32_t z);
 BOUNDS_32 Room_GetWorldBounds(void);
 
 void Room_AlterFloorHeight(ITEM *item, int32_t height);
