+++ src/tr2/game/gun/gun_rifle.c
@@ -7,6 +7,7 @@
 #include "game/lara/misc.h"
 #include "game/output.h"
 #include "game/random.h"
+#include "game/room.h"
 #include "game/sound.h"
 #include "global/const.h"
 #include "global/vars.h"
