+++ src/tr2/game/room.c
@@ -437,34 +437,6 @@ void Room_TestSectorTrigger(const ITEM *const item, const SECTOR *const sector)
     }
 }
 
-int16_t Room_GetIndexFromPos(const int32_t x, const int32_t y, const int32_t z)
-{
-    // TODO: merge this to Room_FindByPos
-    const int32_t room_num = Room_FindByPos(x, y, z);
-    if (room_num == NO_ROOM_NEG) {
-        return NO_ROOM;
-    }
-    return room_num;
-}
-
-int32_t Room_FindByPos(const int32_t x, const int32_t y, const int32_t z)
-{
-    for (int32_t i = 0; i < g_RoomCount; i++) {
-        const ROOM *const room = &g_Rooms[i];
-        const int32_t x1 = room->pos.x + WALL_L;
-        const int32_t x2 = room->pos.x + (room->size.x - 1) * WALL_L;
-        const int32_t y1 = room->max_ceiling;
-        const int32_t y2 = room->min_floor;
-        const int32_t z1 = room->pos.z + WALL_L;
-        const int32_t z2 = room->pos.z + (room->size.z - 1) * WALL_L;
-        if (x >= x1 && x < x2 && y >= y1 && y <= y2 && z >= z1 && z < z2) {
-            return i;
-        }
-    }
-
-    return NO_ROOM_NEG;
-}
-
 int32_t Room_FindGridShift(int32_t src, const int32_t dst)
 {
     const int32_t src_w = src >> WALL_SHIFT;
