+++ src/tr2/game/room.h
@@ -8,8 +8,6 @@
 
 extern int32_t g_FlipEffect;
 
-int16_t Room_GetIndexFromPos(int32_t x, int32_t y, int32_t z);
-int32_t Room_FindByPos(int32_t x, int32_t y, int32_t z);
 int32_t Room_FindGridShift(int32_t src, int32_t dst);
 void Room_GetNearbyRooms(
     int32_t x, int32_t y, int32_t z, int32_t r, int32_t h, int16_t room_num);
